♻️ Refactor homepage into smaller files, readjust youtube embeded size
